Change control/movement method.
Ships now have a "destination", which they will move towards in a straight
line at their maximum speed. This should make it easier to code enemy motion
or simple AI. The player's ship is moved in the same way, by placing the
destination point to the far left or right when the player is pushing left or
right, and the destination on the current position when no direction is pushed.